requirements.txt
Django==1.7.4
django-bower==5.0.4
djangorestframework==2.3.13
djangorestframework-xml==1.0.1
djangorestframework-yaml==1.0.2
django-cors-headers==0.13
gunicorn==19.1.1
futures==2.2.0
trollius==1.0.2
requests==2.4.3
South==0.7.5
MySQL-python==1.2.5
websocket-client==0.21.0
git+https://github.com/avaneesh23/django-admin-tools-stats.git@libraries#egg=django-admin-tools-stats-libraries
